bugfix: timer_observable_generic<RTC> timer stops
The timer_observable_generic<> when instantiated for the RTC
stops receiving interrupts at its rollover condition in both
the app_timer_test and rtc_test applications.

The problem was the calculation of ticks_delta in
timer_observable_generic<>::ticks_update() for non-32 bit
counters.

Fix:
        uint32_t const ticks_delta = (cc_count -
            this->cc_assoc_[cc_index].last_ticks_count_) & counter_mask;

bit-wise and-ing with the counter_mask keeps the value properly in range.