For the SPIS peripheral implement the DMA anomaly 109 work around
http://infocenter.nordicsemi.com/pdf/nRF52_PAN_109_add_v1.1.pdf
System enters IDLE and stops the 64 MHz clock at the same time as the
peripheral that is using DMA is started. This results in the wrong data being
sent to the external device.

Assign the SPI SS pin to a GPIO TE channel which gets triggered by
the falling SS signal.